Remove nuxt util, merge logic to init entry helper
